@@ -20,17 +20,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "20.3.16",
-    "@angular/compiler": "20.3.16",
-    "@angular/core": "20.3.16",
-    "@angular/forms": "20.3.16",
-    "@angular/platform-browser": "20.3.16",
-    "@angular/router": "20.3.16",
+    "@angular/common": "21.1.2",
+    "@angular/compiler": "21.1.2",
+    "@angular/core": "21.1.2",
+    "@angular/forms": "21.1.2",
+    "@angular/platform-browser": "21.1.2",
+    "@angular/router": "21.1.2",
     "postcss": "^8.5.6",
-    "rxjs": "~7.8.0",
-    "tailwindcss": "^3.4.11",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.15.0"
+    "rxjs": "7.8.2",
+    "tailwindcss": "3.4.19",
+    "tslib": "2.8.1",
+    "zone.js": "0.16.0"
   },
   "overrides": {
     "sucrase": {
@@ -38,12 +38,12 @@
     }
   },
   "devDependencies": {
-    "@angular/build": "20.3.15",
-    "@angular/cli": "20.3.15",
-    "@angular/compiler-cli": "20.3.16",
+    "@angular/build": "21.1.2",
+    "@angular/cli": "21.1.2",
+    "@angular/compiler-cli": "21.1.2",
     "@tailwindcss/typography": "^0.5.19",
     "@types/jasmine": "~6.0.0",
-    "autoprefixer": "^10.4.21",
+    "autoprefixer": "10.4.24",
     "jasmine-core": "~6.0.1",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
