@@ -1,5 +1,5 @@
 import { Component, signal } from "@angular/core";
-import { CommonModule } from "@angular/common";
+
 import { FormsModule } from "@angular/forms";
 
 interface Post {
@@ -16,12 +16,12 @@ interface Post {
 @Component({
   selector: "app-social-media-sim",
   standalone: true,
-  imports: [CommonModule, FormsModule],
+  imports: [FormsModule],
   template: `
     <div
       class="min-h-screen py-12 px-4 sm:px-6 lg:px-8"
       style="background-color: #282828;"
-    >
+      >
       <div class="max-w-2xl mx-auto">
         <!-- Header -->
         <div class="mb-8">
@@ -49,13 +49,13 @@ interface Post {
                     fill="none"
                     stroke="currentColor"
                     viewBox="0 0 24 24"
-                  >
+                    >
                     <path
                       stroke-linecap="round"
                       stroke-linejoin="round"
                       stroke-width="2"
                       d="M15.172 7.172a4 4 0 11-5.656 0m0 0l3.5-3.5"
-                    />
+                      />
                   </svg>
                 </button>
                 <button class="btn-secondary">
@@ -64,13 +64,13 @@ interface Post {
                     fill="none"
                     stroke="currentColor"
                     viewBox="0 0 24 24"
-                  >
+                    >
                     <path
                       stroke-linecap="round"
                       stroke-linejoin="round"
                       stroke-width="2"
                       d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z"
-                    />
+                      />
                   </svg>
                 </button>
                 <button
@@ -78,11 +78,11 @@ interface Post {
                   class="btn-primary"
                   [disabled]="!newPostContent().trim()"
                   [class.opacity-50]="!newPostContent().trim()"
-                >
+                  >
                   <svg class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20">
                     <path
                       d="M10.894 2.553a1 1 0 00-1.788 0l-7 14a1 1 0 001.169 1.409l5.951-1.488 5.951 1.488a1 1 0 001.16-1.41l-7-14z"
-                    />
+                      />
                   </svg>
                   Post
                 </button>
@@ -93,98 +93,97 @@ interface Post {
 
         <!-- Feed -->
         <div class="space-y-4">
-          <div *ngFor="let post of posts" class="card p-6 card-hover">
-            <!-- Post Header -->
-            <div class="flex items-start gap-4 mb-4">
-              <div class="text-3xl flex-shrink-0">{{ post.avatar }}</div>
-              <div class="flex-grow min-w-0">
-                <p class="font-bold text-gray-100">{{ post.author }}</p>
-                <p class="text-sm text-gray-500">{{ post.timestamp }}</p>
+          @for (post of posts; track post) {
+            <div class="card p-6 card-hover">
+              <!-- Post Header -->
+              <div class="flex items-start gap-4 mb-4">
+                <div class="text-3xl flex-shrink-0">{{ post.avatar }}</div>
+                <div class="flex-grow min-w-0">
+                  <p class="font-bold text-gray-100">{{ post.author }}</p>
+                  <p class="text-sm text-gray-500">{{ post.timestamp }}</p>
+                </div>
+                <button
+                  class="text-gray-400 hover:text-gray-100 transition-colors"
+                  >
+                  <svg class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20">
+                    <path
+                      d="M6 10a2 2 0 11-4 0 2 2 0 014 0zM12 10a2 2 0 11-4 0 2 2 0 014 0zM16 12a2 2 0 100-4 2 2 0 000 4z"
+                      />
+                  </svg>
+                </button>
               </div>
-              <button
-                class="text-gray-400 hover:text-gray-100 transition-colors"
-              >
-                <svg class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20">
-                  <path
-                    d="M6 10a2 2 0 11-4 0 2 2 0 014 0zM12 10a2 2 0 11-4 0 2 2 0 014 0zM16 12a2 2 0 100-4 2 2 0 000 4z"
-                  />
-                </svg>
-              </button>
-            </div>
-
-            <!-- Post Content -->
-            <p class="text-gray-100 mb-4 leading-relaxed">
-              {{ post.content }}
-            </p>
-
-            <!-- Post Stats -->
-            <div
-              class="flex gap-4 text-sm text-gray-400 mb-4 py-3 border-t border-b border-gray-800"
-            >
-              <span>{{ post.likes }} likes</span>
-              <span>{{ post.comments }} comments</span>
-            </div>
-
-            <!-- Post Actions -->
-            <div class="flex gap-2 justify-around">
-              <button
-                (click)="toggleLike(post.id)"
-                [class.text-primary]="post.isLiked"
-                class="flex-1 flex items-center justify-center gap-2 text-gray-400 hover:text-primary hover:bg-primary/5 py-2 rounded transition-colors duration-250"
-              >
-                <svg
-                  class="w-5 h-5"
-                  [attr.fill]="post.isLiked ? 'currentColor' : 'none'"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
+              <!-- Post Content -->
+              <p class="text-gray-100 mb-4 leading-relaxed">
+                {{ post.content }}
+              </p>
+              <!-- Post Stats -->
+              <div
+                class="flex gap-4 text-sm text-gray-400 mb-4 py-3 border-t border-b border-gray-800"
                 >
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    stroke-width="2"
-                    d="M4.318 6.318a4.5 4.5 0 000 6.364L12 20.364l7.682-7.682a4.5 4.5 0 00-6.364-6.364L12 7.636l-1.318-1.318a4.5 4.5 0 00-6.364 0z"
-                  />
-                </svg>
-                Like
-              </button>
-              <button
-                class="flex-1 flex items-center justify-center gap-2 text-gray-400 hover:text-secondary hover:bg-secondary/5 py-2 rounded transition-colors duration-250"
-              >
-                <svg
-                  class="w-5 h-5"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    stroke-width="2"
-                    d="M8 12h.01M12 12h.01M16 12h.01M21 12c0 4.418-4.03 8-9 8a9.863 9.863 0 01-4.255-.949L3 20l1.395-3.72C3.512 15.042 3 13.574 3 12c0-4.418 4.03-8 9-8s9 3.582 9 8z"
-                  />
-                </svg>
-                Comment
-              </button>
-              <button
-                class="flex-1 flex items-center justify-center gap-2 text-gray-400 hover:text-success hover:bg-success/5 py-2 rounded transition-colors duration-250"
-              >
-                <svg
-                  class="w-5 h-5"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    stroke-width="2"
-                    d="M8.684 13.342C9.589 12.938 10.49 12.5 11.391 12c-.757.689-1.684 1.235-2.707 1.585m0 0a50.991 50.991 0 017.232-12.28m0 0a50.991 50.991 0 015.591 6.347C21.075 8.236 21.5 9.088 21.5 10m0 0H11.5m11.5 0v8.5"
-                  />
-                </svg>
-                Share
-              </button>
+                <span>{{ post.likes }} likes</span>
+                <span>{{ post.comments }} comments</span>
+              </div>
+              <!-- Post Actions -->
+              <div class="flex gap-2 justify-around">
+                <button
+                  (click)="toggleLike(post.id)"
+                  [class.text-primary]="post.isLiked"
+                  class="flex-1 flex items-center justify-center gap-2 text-gray-400 hover:text-primary hover:bg-primary/5 py-2 rounded transition-colors duration-250"
+                  >
+                  <svg
+                    class="w-5 h-5"
+                    [attr.fill]="post.isLiked ? 'currentColor' : 'none'"
+                    stroke="currentColor"
+                    viewBox="0 0 24 24"
+                    >
+                    <path
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                      stroke-width="2"
+                      d="M4.318 6.318a4.5 4.5 0 000 6.364L12 20.364l7.682-7.682a4.5 4.5 0 00-6.364-6.364L12 7.636l-1.318-1.318a4.5 4.5 0 00-6.364 0z"
+                      />
+                  </svg>
+                  Like
+                </button>
+                <button
+                  class="flex-1 flex items-center justify-center gap-2 text-gray-400 hover:text-secondary hover:bg-secondary/5 py-2 rounded transition-colors duration-250"
+                  >
+                  <svg
+                    class="w-5 h-5"
+                    fill="none"
+                    stroke="currentColor"
+                    viewBox="0 0 24 24"
+                    >
+                    <path
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                      stroke-width="2"
+                      d="M8 12h.01M12 12h.01M16 12h.01M21 12c0 4.418-4.03 8-9 8a9.863 9.863 0 01-4.255-.949L3 20l1.395-3.72C3.512 15.042 3 13.574 3 12c0-4.418 4.03-8 9-8s9 3.582 9 8z"
+                      />
+                  </svg>
+                  Comment
+                </button>
+                <button
+                  class="flex-1 flex items-center justify-center gap-2 text-gray-400 hover:text-success hover:bg-success/5 py-2 rounded transition-colors duration-250"
+                  >
+                  <svg
+                    class="w-5 h-5"
+                    fill="none"
+                    stroke="currentColor"
+                    viewBox="0 0 24 24"
+                    >
+                    <path
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                      stroke-width="2"
+                      d="M8.684 13.342C9.589 12.938 10.49 12.5 11.391 12c-.757.689-1.684 1.235-2.707 1.585m0 0a50.991 50.991 0 017.232-12.28m0 0a50.991 50.991 0 015.591 6.347C21.075 8.236 21.5 9.088 21.5 10m0 0H11.5m11.5 0v8.5"
+                      />
+                  </svg>
+                  Share
+                </button>
+              </div>
             </div>
-          </div>
+          }
 
           <!-- Load More -->
           <button (click)="loadMorePosts()" class="w-full btn-secondary py-3">
@@ -193,7 +192,7 @@ interface Post {
         </div>
       </div>
     </div>
-  `,
+    `,
 })
 export class SocialMediaSimComponent {
   newPostContent = signal("");
